# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ARM-112 Training Service API
  version: 0.3.7
  description: |
    Контракт учебного сервиса подготовки операторов ДДС: три роли (администратор, преподаватель,
    обучающийся), два режима занятий (заполнение карточки оператора 112 и действия с карточкой ДДС),
    персональные занятия, оценка ИИ с приоритетом оценки преподавателя, аудит.
    Все ответы строгие (`additionalProperties: false`): новое поле — новая версия контракта.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
    contractVersion: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      security:
        - contractVersion: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Вход выполнен
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '429': {$ref: '#/components/responses/ErrorResponse'}
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      security:
        - contractVersion: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefreshRequest'}
      responses:
        '200':
          description: Токены ротированы; переданный refresh-токен отозван
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      security:
        - contractVersion: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LogoutRequest'}
      responses:
        '204': {description: Refresh-токен отозван}
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      responses:
        '200':
          description: Текущий пользователь
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
  /auth/password:
    post:
      tags: [Auth]
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PasswordChangeRequest'}
      responses:
        '204': {description: Пароль изменён}
        '422': {$ref: '#/components/responses/ErrorResponse'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    contractVersion:
      type: apiKey
      in: header
      name: X-Contract-Version
      description: Версия контракта клиента. Для текущего API укажите 0.3.
...
  responses:
    ErrorResponse:
      description: Ошибка
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
...
    Role:
      type: string
      enum: [ADMIN, TEACHER, TRAINEE]
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, minLength: 1, maxLength: 100}
        password: {type: string, minLength: 1, maxLength: 200}
    AuthResponse:
      type: object
      additionalProperties: false
      required: [accessToken, expiresAt, user, refreshToken, refreshExpiresAt]
      properties:
        accessToken: {type: string}
        expiresAt: {type: string, format: date-time}
        user: {$ref: '#/components/schemas/User'}
        refreshToken: {type: string, description: Одноразовый токен для ротации пары токенов}
        refreshExpiresAt: {type: string, format: date-time}
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken: {type: string, minLength: 1}
    LogoutRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken: {type: string, minLength: 1}
    PasswordChangeRequest:
      type: object
      additionalProperties: false
      required: [current, next]
      properties:
        current: {type: string}
        next: {type: string, minLength: 4, maxLength: 200}
...
    User:
      type: object
      additionalProperties: false
      required: [id, displayName, role, login, workstationNumber, groupId]
      properties:
        id: {type: string, format: uuid}
        displayName: {type: string}
        role: {$ref: '#/components/schemas/Role'}
        login: {type: string}
        workstationNumber: {type: [string, 'null'], description: "Номер АРМ обучающегося из учётной записи"}
        groupId: {type: [string, 'null'], format: uuid}
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId, fieldErrors, details]
          properties:
            code: {type: string}
            message: {type: string}
            requestId: {type: string, format: uuid}
            fieldErrors:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [path, code, message]
                properties:
                  path: {type: string}
                  code: {type: string}
                  message: {type: string}
            details: {type: object, additionalProperties: true}